IdGenerater代码重构
